+++ src/minisphere/game.h
@@ -62,7 +62,7 @@ const char*      game_save_id        (const game_t* it);
 const path_t*    game_script_path    (const game_t* it);
 const char*      game_summary        (const game_t* it);
 int              game_version        (const game_t* it);
-path_t*          game_build_path     (const game_t* it, const char* filename, const char* base_dir_name, bool legacy);
+path_t*          game_build_path     (const game_t* it, const char* filename, const char* base_dir_name, bool legacy_mode);
 bool             game_mkdir          (game_t* it, const char* dirname);
 void*            game_read_file      (game_t* it, const char* filename, size_t *out_size);
 bool             game_rename         (game_t* it, const char* filename1, const char* filename2);